Move lib/client/BaseGraphqlLauncher.js below lib/client/graphql/
