+++ app/graphql/clients/BaseAppGraphqlLauncher.js
@@ -1,4 +1,4 @@
-import BaseGraphqlLauncher from '../../../lib/client/BaseGraphqlLauncher.js'
+import BaseGraphqlLauncher from '../../../lib/client/graphql/BaseGraphqlLauncher.js'
 
 import graphqlConfig from '../graphql.config.js'
