+++ lib/client/graphql/BaseGraphqlLauncher.js
@@ -1,4 +1,4 @@
-import ProgressHttpFetcher from '../tools/ProgressHttpFetcher.js'
+import ProgressHttpFetcher from '../../tools/ProgressHttpFetcher.js'
 
 /**
  * Base class of GraphQL launcher.
@@ -457,7 +457,7 @@ export default class BaseGraphqlLauncher {
    *
    * @param {{
    *   payload: furo.Payload<*>
-   *   sink?: Partial<import('../tools/ProgressHttpFetcher').ProgressHttpFetcherSink>
+   *   sink?: Partial<import('../../tools/ProgressHttpFetcher.js').ProgressHttpFetcherSink>
    * }} params - Parameters.
    * @returns {Promise<Response | null>} Instance of fetch API response.
    */
@@ -575,7 +575,7 @@ export default class BaseGraphqlLauncher {
  *   rawResponse: Response | null
  *   payload: furo.Payload<*> | null
  *   result: object | null
- *   abortedReason?: import('./graphql/BaseGraphqlCapsule.js').LAUNCH_ABORTED_REASON
+ *   abortedReason?: import('./BaseGraphqlCapsule.js').LAUNCH_ABORTED_REASON
  * }} CapsuleParams
  */
